requirements.txt
openpyxl
fuzzywuzzy
python-magic
python-magic-bin
libmagic
xlrd